Scala imports cleanup.
