+++ src/main/scala/io/archivesunleashed/app/CommandLineApp.scala
@@ -23,8 +23,8 @@ import org.apache.log4j.Logger
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{Dataset, Row}
 import org.apache.spark.{SparkConf, SparkContext}
-import org.rogach.scallop.ScallopConf
 import org.rogach.scallop.exceptions.ScallopException
+import org.rogach.scallop.ScallopConf
 
 /* Usage:
  *
