+++ src/main/scala/io/archivesunleashed/matchbox/DetectMimeTypeTika.scala
@@ -18,11 +18,11 @@ package io.archivesunleashed.matchbox
 // scalastyle:off underscore.import
 import scala.collection.JavaConverters._
 // scalastyle:on underscore.import
-import org.apache.tika.Tika
 import org.apache.tika.detect.DefaultDetector
 import org.apache.tika.io.TikaInputStream
 import org.apache.tika.mime.MimeTypes
 import org.apache.tika.parser.AutoDetectParser
+import org.apache.tika.Tika
 
 /** Detect MIME type using Apache Tika. */
 object DetectMimeTypeTika {
