+++ src/test/scala/io/archivesunleashed/df/SimpleDfTest.scala
@@ -18,8 +18,8 @@ package io.archivesunleashed
 
 import com.google.common.io.Resources
 import io.archivesunleashed.df.ExtractDomainDF
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.functions.desc
+import org.apache.spark.sql.SparkSession
 import org.apache.spark.{SparkConf, SparkContext}
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
